Front Update

+++ client/src/hooks/useFavoriteHotel.js
@@ -1,7 +1,7 @@
 import axios from "axios";
 import Cookies from 'js-cookie';
-import { URL_origin } from "../utils/helpers.js"
-import { useCallback, useEffect, useMemo, useState, memo } from "react"
+import { URL_origin } from "../utils/helpers.js";
+import { useCallback, useEffect, useState } from "react";
 
 const useFavoriteHotel = ({ hotelId, userId = null }) => {
   const [isFavorite, setIsFavorite] = useState(false);
@@ -17,9 +17,9 @@ const useFavoriteHotel = ({ hotelId, userId = null }) => {
       }
     }).then((res) => {
       if(res.data.hotel) {
-        setIsFavorite(true)
+        setIsFavorite(true);
       } else {
-        setIsFavorite(false)
+        setIsFavorite(false);
       }
     }).catch(err => {
       console.log(err);
@@ -27,14 +27,18 @@ const useFavoriteHotel = ({ hotelId, userId = null }) => {
   }, []);
 
   const toggleFavorite = useCallback(() => {
+    if(userId == null) {
+      return;
+    }
     if(isFavorite) {
       axios.delete(`${URL_origin}/api/client/favorite/${hotelId}`, {
         headers: {
           Authorization: `Bearer ${accessToken}`
         }
       }).then((res) => {
+        console.log(res);
         if(res.status === 201) {
-          setIsFavorite(false)
+          setIsFavorite(false);
         }
       }).catch(err => {
         console.log(err);
@@ -48,6 +52,7 @@ const useFavoriteHotel = ({ hotelId, userId = null }) => {
           Authorization: `Bearer ${accessToken}`
         }
       }).then((res) => {
+        console.log(res);
         if(res.status == 201) {
           setIsFavorite(true);
         }
